adding initial test class for SQL Data Provider
